requirements.txt
GitPython==3.1.29
pyserial==3.5
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
slack-bolt==1.15.3
slack-sdk==3.19.3
